+++ src/main/java/Models/Parameters.java
@@ -3,13 +3,14 @@ package Models;
 public class Parameters {
     public Integer total_agents;
     public Integer total_steps;
-    public Integer space_width;
+    public Double space_width;
     public Double velocity;
     public Double radius;
     public Double theta_amp;
     public String folder;
     public Double runs;
-    public Parameters(Integer total_agents, Integer total_steps, Integer space_width, Double velocity, Double radius, Double theta_amp, String folder , Double runs) {
+
+    public Parameters(Integer total_agents, Integer total_steps, Double space_width, Double velocity, Double radius, Double theta_amp, String folder , Double runs) {
         this.total_agents = total_agents;
         this.total_steps = total_steps;
         this.space_width = space_width;
